Usar BigDecimal para todos los importes en lugar de double
double es binario y no representa exactamente valores como 0.10 o 89.90.
En una tienda eso se traduce en totales con arrastre (59.970000000000006)
que acaban impresos en la factura y guardados en el saldo de la tarjeta.

Se convierten a BigDecimal los diez campos monetarios del modelo
(precio, total, precio unitario, saldo de tarjeta y de la tienda, costo
de envio, monto del pago y los tres importes de la factura), los DTOs
que los transportan y las cinco consultas de reportes.

La aritmetica de PedidoService pasa a add/subtract/multiply. La
comparacion de saldo usa compareTo y no equals: equals distingue 10.0 de
10.00 porque tiene en cuenta la escala, y aqui interesa el valor.

Las entidades declaran @Column(precision = 12, scale = 2), que es lo que
usara Hibernate al crear la tabla en una base nueva.

FALTA UN PASO MANUAL EN LA BASE DE DATOS EXISTENTE

La aplicacion arranca con ddl-auto=update, y ese modo solo añade tablas
y columnas que falten: nunca cambia el tipo de una columna que ya existe.
Sin ejecutar migraciones/001-importes-a-decimal.sql los importes se
siguen guardando en columnas DOUBLE y el arreglo queda a medias: exacto
al calcular, inexacto al guardar. Se nota ya en la API, que devuelve
99.9 en vez de 99.90 porque el valor viene de una columna DOUBLE.

Pruebas: se añade Importes con los helpers imp() y assertImporte(), que
compara por valor y no por escala, y dos casos nuevos que fijan la
precision: 19.99 x 3 mas 10.01 de envio da exactamente 69.98, y 0.10
diez veces da exactamente 1.00. Ambos fallaban con double.

Verificado: 134 pruebas y la aplicacion arranca contra MySQL sin errores.
El JSON no cambia de forma, asi que el frontend no se ve afectado.

+++ BackendTienda/src/main/java/com/backend/AltaPinta/dto/CarritoDTO.java
@@ -1,11 +1,13 @@
 package com.backend.AltaPinta.dto;
 
+import java.math.BigDecimal;
+
 import java.util.List;
 
 public class CarritoDTO {
 
     private Long id;
-    private Double total;
+    private BigDecimal total;
     private List<CarritoItemDTO> items;
 
     public Long getId() {
@@ -16,11 +18,11 @@ public class CarritoDTO {
         this.id = id;
     }
 
-    public Double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
