+++ BackendTienda/src/main/java/com/backend/AltaPinta/dto/TarjetaResponse.java
@@ -1,9 +1,11 @@
 package com.backend.AltaPinta.dto;
 
+import java.math.BigDecimal;
+
 public record TarjetaResponse(
         Long id,
         String numero,
         String titular,
-        Double saldo
+        BigDecimal saldo
 ) {}
